Reorganize `optimizations`
