+++ creusot/src/backend/optimization/invariants.rs
@@ -26,7 +26,7 @@ use rustc_middle::{
 use rustc_span::DUMMY_SP;
 
 /// Add loop invariants to `body` for each mutable borrow that is _not_ modified in a loop.
-pub fn infer_proph_invariants<'tcx>(
+pub(crate) fn infer_proph_invariants<'tcx>(
     ctx: &TranslationCtx<'tcx>,
     body: &mut fmir::Body<'tcx>,
     body_id: BodyId,
@@ -51,7 +51,7 @@ pub fn infer_proph_invariants<'tcx>(
 
             let local = Ident::fresh_local(format!("old_{}_{ix}", k.as_u32()));
             let subst = ctx.mk_args(&[u.ty(tcx, &body.locals).into()]);
-            let ty = Ty::new_adt(ctx.tcx, ctx.adt_def(snap_ty), subst);
+            let ty = Ty::new_adt(tcx, ctx.adt_def(snap_ty), subst);
 
             body.locals.insert(local, fmir::LocalDecl {
                 span: DUMMY_SP,
